Refactor http-auth into client-http
