+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/auth/HttpApiKeyAuthSigner.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.runtime.http.auth;
+package software.amazon.smithy.java.runtime.client.http.auth;
 
 import java.net.http.HttpHeaders;
 import java.util.LinkedHashMap;
