+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/auth/HttpBasicAuthAuthScheme.java
@@ -3,24 +3,25 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.runtime.http.auth;
+package software.amazon.smithy.java.runtime.client.http.auth;
 
 import software.amazon.smithy.java.runtime.auth.api.Signer;
-import software.amazon.smithy.java.runtime.auth.api.identity.TokenIdentity;
+import software.amazon.smithy.java.runtime.auth.api.identity.LoginIdentity;
 import software.amazon.smithy.java.runtime.client.core.auth.scheme.AuthScheme;
 import software.amazon.smithy.java.runtime.client.core.auth.scheme.AuthSchemeFactory;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpRequest;
 import software.amazon.smithy.model.shapes.ShapeId;
-import software.amazon.smithy.model.traits.HttpBearerAuthTrait;
+import software.amazon.smithy.model.traits.HttpBasicAuthTrait;
 
 /**
- * Implements the HTTP Bearer Authentication Scheme as defined in
+ * Implements the HTTP Basic Authentication Scheme as defined in
  * <a href="https://datatracker.ietf.org/doc/html/rfc2617.html">RFC 2617</a>.
  */
-public final class HttpBearerAuthScheme implements AuthScheme<SmithyHttpRequest, TokenIdentity> {
+public final class HttpBasicAuthAuthScheme implements AuthScheme<SmithyHttpRequest, LoginIdentity> {
+
     @Override
     public ShapeId schemeId() {
-        return HttpBearerAuthTrait.ID;
+        return HttpBasicAuthTrait.ID;
     }
 
     @Override
@@ -29,25 +30,25 @@ public final class HttpBearerAuthScheme implements AuthScheme<SmithyHttpRequest,
     }
 
     @Override
-    public Class<TokenIdentity> identityClass() {
-        return TokenIdentity.class;
+    public Class<LoginIdentity> identityClass() {
+        return LoginIdentity.class;
     }
 
     @Override
-    public Signer<SmithyHttpRequest, TokenIdentity> signer() {
-        return HttpBearerAuthSigner.INSTANCE;
+    public Signer<SmithyHttpRequest, LoginIdentity> signer() {
+        return HttpBasicAuthSigner.INSTANCE;
     }
 
-    public static final class Factory implements AuthSchemeFactory<HttpBearerAuthTrait> {
+    public static final class Factory implements AuthSchemeFactory<HttpBasicAuthTrait> {
 
         @Override
         public ShapeId schemeId() {
-            return HttpBearerAuthTrait.ID;
+            return HttpBasicAuthTrait.ID;
         }
 
         @Override
-        public AuthScheme<?, ?> createAuthScheme(HttpBearerAuthTrait trait) {
-            return new HttpBearerAuthScheme();
+        public AuthScheme<?, ?> createAuthScheme(HttpBasicAuthTrait trait) {
+            return new HttpBasicAuthAuthScheme();
         }
     }
 }
