+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/auth/HttpBearerAuthSigner.java
@@ -3,38 +3,34 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.runtime.http.auth;
+package software.amazon.smithy.java.runtime.client.http.auth;
 
 import java.net.http.HttpHeaders;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import software.amazon.smithy.java.logging.InternalLogger;
 import software.amazon.smithy.java.runtime.auth.api.AuthProperties;
 import software.amazon.smithy.java.runtime.auth.api.Signer;
-import software.amazon.smithy.java.runtime.auth.api.identity.LoginIdentity;
+import software.amazon.smithy.java.runtime.auth.api.identity.TokenIdentity;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpRequest;
 
-final class HttpBasicAuthSigner implements Signer<SmithyHttpRequest, LoginIdentity> {
-    static final HttpBasicAuthSigner INSTANCE = new HttpBasicAuthSigner();
-    private static final InternalLogger LOGGER = InternalLogger.getLogger(HttpBasicAuthSigner.class);
+final class HttpBearerAuthSigner implements Signer<SmithyHttpRequest, TokenIdentity> {
+    static final HttpBearerAuthSigner INSTANCE = new HttpBearerAuthSigner();
+    private static final InternalLogger LOGGER = InternalLogger.getLogger(HttpBearerAuthSigner.class);
     private static final String AUTHORIZATION_HEADER = "Authorization";
-    private static final String SCHEME = "Basic";
+    private static final String SCHEME = "Bearer";
 
-    private HttpBasicAuthSigner() {}
+    private HttpBearerAuthSigner() {}
 
     @Override
     public CompletableFuture<SmithyHttpRequest> sign(
         SmithyHttpRequest request,
-        LoginIdentity identity,
+        TokenIdentity identity,
         AuthProperties properties
     ) {
-        var identityString = identity.username() + ":" + identity.password();
-        var base64Value = Base64.getEncoder().encodeToString(identityString.getBytes(StandardCharsets.UTF_8));
         var headers = new LinkedHashMap<>(request.headers().map());
-        var existing = headers.put(AUTHORIZATION_HEADER, List.of(SCHEME + " " + base64Value));
+        var existing = headers.put(AUTHORIZATION_HEADER, List.of(SCHEME + " " + identity.token()));
         if (existing != null) {
             LOGGER.debug("Replaced existing Authorization header value.");
         }
