+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/auth/HttpDigestAuthAuthScheme.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.runtime.http.auth;
+package software.amazon.smithy.java.runtime.client.http.auth;
 
 import software.amazon.smithy.java.runtime.auth.api.Signer;
 import software.amazon.smithy.java.runtime.auth.api.identity.LoginIdentity;
@@ -11,17 +11,17 @@ import software.amazon.smithy.java.runtime.client.core.auth.scheme.AuthScheme;
 import software.amazon.smithy.java.runtime.client.core.auth.scheme.AuthSchemeFactory;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpRequest;
 import software.amazon.smithy.model.shapes.ShapeId;
-import software.amazon.smithy.model.traits.HttpBasicAuthTrait;
+import software.amazon.smithy.model.traits.HttpDigestAuthTrait;
 
 /**
- * Implements the HTTP Basic Authentication Scheme as defined in
+ * Implements the HTTP Digest Authentication Scheme as defined in
  * <a href="https://datatracker.ietf.org/doc/html/rfc2617.html">RFC 2617</a>.
  */
-public final class HttpBasicAuthAuthScheme implements AuthScheme<SmithyHttpRequest, LoginIdentity> {
+public final class HttpDigestAuthAuthScheme implements AuthScheme<SmithyHttpRequest, LoginIdentity> {
 
     @Override
     public ShapeId schemeId() {
-        return HttpBasicAuthTrait.ID;
+        return HttpDigestAuthTrait.ID;
     }
 
     @Override
@@ -36,19 +36,19 @@ public final class HttpBasicAuthAuthScheme implements AuthScheme<SmithyHttpReque
 
     @Override
     public Signer<SmithyHttpRequest, LoginIdentity> signer() {
-        return HttpBasicAuthSigner.INSTANCE;
+        return HttpDigestAuthSigner.INSTANCE;
     }
 
-    public static final class Factory implements AuthSchemeFactory<HttpBasicAuthTrait> {
+    public static final class Factory implements AuthSchemeFactory<HttpDigestAuthTrait> {
 
         @Override
         public ShapeId schemeId() {
-            return HttpBasicAuthTrait.ID;
+            return HttpDigestAuthTrait.ID;
         }
 
         @Override
-        public AuthScheme<?, ?> createAuthScheme(HttpBasicAuthTrait trait) {
-            return new HttpBasicAuthAuthScheme();
+        public AuthScheme<?, ?> createAuthScheme(HttpDigestAuthTrait trait) {
+            return new HttpDigestAuthAuthScheme();
         }
     }
 }
