+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/auth/HttpDigestAuthSigner.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.runtime.http.auth;
+package software.amazon.smithy.java.runtime.client.http.auth;
 
 import java.util.concurrent.CompletableFuture;
 import software.amazon.smithy.java.runtime.auth.api.AuthProperties;
