+++ codegen/client/src/it/resources/META-INF/smithy/manifest
@@ -1,1 +1,2 @@
 main.smithy
+test-auth-scheme.smithy
