+++ codegen/client/src/it/resources/META-INF/smithy/test-auth-scheme.smithy
@@ -1,0 +1,7 @@
+$version: "2"
+
+namespace smithy.test.auth
+
+@trait
+@authDefinition
+structure testAuthScheme {}
