+++ codegen/client/src/test/resources/META-INF/services/software.amazon.smithy.model.traits.TraitService
@@ -1,0 +1,1 @@
+software.amazon.smithy.java.codegen.client.TestAuthSchemeTrait$Provider
